dataset/acquisition/screenshot_processing/requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
language-tool-python>=2.9.4
